requirements.txt
Flask==2.3.2
Flask_Login==0.6.2
python-dotenv==1.0.0
python_bcrypt==0.3.2
Requests==2.31.0

bcrypt~=4.0.1
SQLAlchemy~=2.0.20